Переносит код отрисовки миниатюр из main.js в rendering-of-thumbnails.js
