+++ js/main.js
@@ -1,26 +1,17 @@
-import {showthumbnail} from './rendering-of-thumbnails.js';
-import {showBigImage, showPostPreview} from './rendering-of-big-image.js';
-import {loadData} from './server-load.js';
 import './form.js';
+import {loadData} from './server-load.js';
+import {getImageWrapper} from './rendering-of-thumbnails.js';
+import {showErrorLoadWrapper} from './util.js';
 
-const imagesWrapper = document.querySelector('.pictures');
-
-const showUserPost = (optionImage) => {
-  showBigImage();
-  showPostPreview(optionImage);
-};
+const imageFilters = document.querySelector('.img-filters');
 
 loadData()
   .then((image) => {
-    image.forEach((item) => {
-      const thumbnailElement = showthumbnail(item);
-      imagesWrapper.append(thumbnailElement);
-    });
-    imagesWrapper.addEventListener('click', (evt) => {
-      const optionImage = image.find((element) => element.id === +evt.target.dataset.id);
-      if(!optionImage) {
-        return;
-      }
-      showUserPost(optionImage);
-    });
+    if (image) {
+      getImageWrapper(image);
+    }
+  }).catch((err) => {
+    showErrorLoadWrapper(`Ошибка получения данных с сервера. ${err}`);
   });
+
+imageFilters.classList.remove('img-filters--inactive');
